-добавляет обработку ошибок в компоненты -удаляет неиспользуемый код -переносит общий стэйт
