--- src/mocks/details.ts
@@ -1,191 +1,0 @@
-import { TDetail } from '../types/details';
-
-const DETAILS: TDetail[] = [
-  {
-    id: '372dcd05-d179-4a2a-be70-0b08daada439',
-    title: 'Nice, cozy, warm big bed apartment',
-    description:
-      'Relax, rejuvenate and unplug in this ultimate rustic getaway experience in the country. In our beautiful screened Pondhouse, you can gaze at the stars and listen to the sounds of nature from your cozy warm bed.',
-    type: 'room',
-    price: 167,
-    images: [
-      'https://13.design.pages.academy/static/hotel/18.jpg',
-      'https://13.design.pages.academy/static/hotel/11.jpg',
-      'https://13.design.pages.academy/static/hotel/20.jpg',
-      'https://13.design.pages.academy/static/hotel/8.jpg',
-      'https://13.design.pages.academy/static/hotel/10.jpg',
-      'https://13.design.pages.academy/static/hotel/9.jpg',
-    ],
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36554,
-      longitude: 4.911976,
-      zoom: 16,
-    },
-    goods: [
-      'Dishwasher',
-      'Washing machine',
-      'Laptop friendly workspace',
-      'Air conditioning',
-      'Towels',
-      'Breakfast',
-      'Washer',
-      'Kitchen',
-      'Cable TV',
-      'Baby seat',
-      'Coffee machine',
-      'Fridge',
-    ],
-    host: {
-      isPro: true,
-      name: 'Angelina',
-      avatarUrl:
-        'https://13.design.pages.academy/static/host/avatar-angelina.jpg',
-    },
-    isPremium: true,
-    isFavorite: false,
-    rating: 3.4,
-    bedrooms: 1,
-    maxAdults: 1,
-  },
-  {
-    id: 'f21b1d07-4ed8-456f-92ac-fd13b5cf6c19',
-    title: 'Loft Studio in the Central Area',
-    description:
-      'Relax, rejuvenate and unplug in this ultimate rustic getaway experience in the country. In our beautiful screened Pondhouse, you can gaze at the stars and listen to the sounds of nature from your cozy warm bed.',
-    type: 'room',
-    price: 298,
-    images: [
-      'https://13.design.pages.academy/static/hotel/14.jpg',
-      'https://13.design.pages.academy/static/hotel/2.jpg',
-      'https://13.design.pages.academy/static/hotel/19.jpg',
-      'https://13.design.pages.academy/static/hotel/20.jpg',
-      'https://13.design.pages.academy/static/hotel/16.jpg',
-      'https://13.design.pages.academy/static/hotel/18.jpg',
-    ],
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.397540000000006,
-      longitude: 4.9099759999999995,
-      zoom: 16,
-    },
-    goods: ['Wi-Fi', 'Washing machine', 'Cable TV', 'Heating'],
-    host: {
-      isPro: true,
-      name: 'Angelina',
-      avatarUrl:
-        'https://13.design.pages.academy/static/host/avatar-angelina.jpg',
-    },
-    isPremium: true,
-    isFavorite: false,
-    rating: 3.6,
-    bedrooms: 1,
-    maxAdults: 1,
-  },
-  {
-    id: 'f21b1d07-4ed8-456f-92ac-fd13b5cf6c19',
-    title: 'Loft Studio in the Central Area',
-    description:
-      'Relax, rejuvenate and unplug in this ultimate rustic getaway experience in the country. In our beautiful screened Pondhouse, you can gaze at the stars and listen to the sounds of nature from your cozy warm bed.',
-    type: 'room',
-    price: 298,
-    images: [
-      'https://13.design.pages.academy/static/hotel/14.jpg',
-      'https://13.design.pages.academy/static/hotel/2.jpg',
-      'https://13.design.pages.academy/static/hotel/19.jpg',
-      'https://13.design.pages.academy/static/hotel/20.jpg',
-      'https://13.design.pages.academy/static/hotel/16.jpg',
-      'https://13.design.pages.academy/static/hotel/18.jpg',
-    ],
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.397540000000006,
-      longitude: 4.9099759999999995,
-      zoom: 16,
-    },
-    goods: ['Wi-Fi', 'Washing machine', 'Cable TV', 'Heating'],
-    host: {
-      isPro: true,
-      name: 'Angelina',
-      avatarUrl:
-        'https://13.design.pages.academy/static/host/avatar-angelina.jpg',
-    },
-    isPremium: true,
-    isFavorite: false,
-    rating: 3.6,
-    bedrooms: 1,
-    maxAdults: 1,
-  },
-  {
-    id: '520aab03-3fbb-4528-8903-8f373247982b',
-    title: 'Tile House',
-    description:
-      'Relax, rejuvenate and unplug in this ultimate rustic getaway experience in the country. In our beautiful screened Pondhouse, you can gaze at the stars and listen to the sounds of nature from your cozy warm bed.',
-    type: 'house',
-    price: 898,
-    images: [
-      'https://13.design.pages.academy/static/hotel/15.jpg',
-      'https://13.design.pages.academy/static/hotel/13.jpg',
-      'https://13.design.pages.academy/static/hotel/2.jpg',
-      'https://13.design.pages.academy/static/hotel/7.jpg',
-      'https://13.design.pages.academy/static/hotel/17.jpg',
-      'https://13.design.pages.academy/static/hotel/1.jpg',
-    ],
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.37454,
-      longitude: 4.881976,
-      zoom: 16,
-    },
-    goods: [
-      'Dishwasher',
-      'Breakfast',
-      'Coffee machine',
-      'Baby seat',
-      'Fridge',
-      'Washing machine',
-      'Kitchen',
-    ],
-    host: {
-      isPro: true,
-      name: 'Angelina',
-      avatarUrl:
-        'https://13.design.pages.academy/static/host/avatar-angelina.jpg',
-    },
-    isPremium: false,
-    isFavorite: false,
-    rating: 1.9,
-    bedrooms: 1,
-    maxAdults: 2,
-  },
-];
-
-export { DETAILS };
