--- src/mocks/near-places.ts
@@ -1,443 +1,0 @@
-import { TOffer } from '../types/offer';
-
-const NEAR_PLACES: TOffer[] = [
-  {
-    id: '768226ff-11f3-4bec-b13c-a903de1da7de',
-    title: 'Perfectly located Castro',
-    type: 'room',
-    price: 253,
-    previewImage: 'https://13.design.pages.academy/static/hotel/9.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.385540000000006,
-      longitude: 4.902976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.4,
-  },
-  {
-    id: 'f21b1d07-4ed8-456f-92ac-fd13b5cf6c19',
-    title: 'Loft Studio in the Central Area',
-    type: 'room',
-    price: 298,
-    previewImage: 'https://13.design.pages.academy/static/hotel/20.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.397540000000006,
-      longitude: 4.9099759999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.6,
-  },
-  {
-    id: '520aab03-3fbb-4528-8903-8f373247982b',
-    title: 'Tile House',
-    type: 'house',
-    price: 898,
-    previewImage: 'https://13.design.pages.academy/static/hotel/8.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.37454,
-      longitude: 4.881976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.9,
-  },
-  {
-    id: '15855756-3589-4f06-b2cf-9da06fb87720',
-    title: 'Tile House',
-    type: 'house',
-    price: 771,
-    previewImage: 'https://13.design.pages.academy/static/hotel/7.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.367540000000005,
-      longitude: 4.883976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.3,
-  },
-  {
-    id: '60695df5-8839-4b34-a6db-6cb634d5cce3',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'room',
-    price: 165,
-    previewImage: 'https://13.design.pages.academy/static/hotel/16.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.361540000000005,
-      longitude: 4.883976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4,
-  },
-  {
-    id: 'a92c6c59-054a-4ae1-a424-fd18db9eac54',
-    title: 'The Pondhouse - A Magical Place',
-    type: 'house',
-    price: 781,
-    previewImage: 'https://13.design.pages.academy/static/hotel/7.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.35754,
-      longitude: 4.9179759999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4,
-  },
-  {
-    id: '66107393-5ba9-4ca2-83ae-657ff48e4aed',
-    title: 'The Pondhouse - A Magical Place',
-    type: 'house',
-    price: 141,
-    previewImage: 'https://13.design.pages.academy/static/hotel/6.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.37854,
-      longitude: 4.894976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.7,
-  },
-  {
-    id: 'cb19e9e4-5abc-4464-bd7c-48a50b4d2325',
-    title: 'Loft Studio in the Central Area',
-    type: 'house',
-    price: 218,
-    previewImage: 'https://13.design.pages.academy/static/hotel/11.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36854,
-      longitude: 4.887976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.9,
-  },
-  {
-    id: '6537145f-0166-4faa-877c-12bd1efff655',
-    title: 'Wood and stone place',
-    type: 'room',
-    price: 294,
-    previewImage: 'https://13.design.pages.academy/static/hotel/2.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.370540000000005,
-      longitude: 4.9099759999999995,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.7,
-  },
-  {
-    id: '8dbd3d28-f4e5-4894-94dd-e64a461c390e',
-    title: 'The Joshua Tree House',
-    type: 'room',
-    price: 106,
-    previewImage: 'https://13.design.pages.academy/static/hotel/3.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.389540000000004,
-      longitude: 4.883976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.2,
-  },
-  {
-    id: '40a275c9-ce93-4892-ac53-ba221ba0deb6',
-    title: 'The Joshua Tree House',
-    type: 'hotel',
-    price: 254,
-    previewImage: 'https://13.design.pages.academy/static/hotel/13.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.35054,
-      longitude: 4.908976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.9,
-  },
-  {
-    id: '0a329088-297e-41eb-ab31-23993cb08985',
-    title: 'Perfectly located Castro',
-    type: 'room',
-    price: 173,
-    previewImage: 'https://13.design.pages.academy/static/hotel/6.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.37154,
-      longitude: 4.889976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 5,
-  },
-  {
-    id: 'c91391dc-9b8a-401e-bc8b-21ceb1ab7747',
-    title: 'House in countryside',
-    type: 'room',
-    price: 117,
-    previewImage: 'https://13.design.pages.academy/static/hotel/10.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.364540000000005,
-      longitude: 4.9019759999999994,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.2,
-  },
-  {
-    id: '00c35338-7006-4fdd-adf7-8cdeb0c7df0a',
-    title: 'Canal View Prinsengracht',
-    type: 'house',
-    price: 275,
-    previewImage: 'https://13.design.pages.academy/static/hotel/15.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36354,
-      longitude: 4.911976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.5,
-  },
-  {
-    id: '444b33e7-4abe-4fdc-84e3-f8f4f083c76e',
-    title: 'Waterfront with extraordinary view',
-    type: 'house',
-    price: 941,
-    previewImage: 'https://13.design.pages.academy/static/hotel/16.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36354,
-      longitude: 4.889976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.8,
-  },
-  {
-    id: 'd32e370e-3f11-4573-899d-b2cb3a09d473',
-    title: 'Waterfront with extraordinary view',
-    type: 'house',
-    price: 708,
-    previewImage: 'https://13.design.pages.academy/static/hotel/5.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.388540000000006,
-      longitude: 4.899976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.5,
-  },
-  {
-    id: 'f222c238-b78d-498e-83e9-258d0822c818',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'room',
-    price: 235,
-    previewImage: 'https://13.design.pages.academy/static/hotel/10.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36954,
-      longitude: 4.914976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.1,
-  },
-  {
-    id: '72f118b5-aa01-40c6-a917-050a3faa2a06',
-    title: 'Beautiful & luxurious apartment at great location',
-    type: 'apartment',
-    price: 312,
-    previewImage: 'https://13.design.pages.academy/static/hotel/16.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.37554,
-      longitude: 4.9019759999999994,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3.3,
-  },
-  {
-    id: '2bcad854-b0df-46ff-aa10-eb83b292737a',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'room',
-    price: 250,
-    previewImage: 'https://13.design.pages.academy/static/hotel/2.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.385540000000006,
-      longitude: 4.886976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.1,
-  },
-];
-
-export { NEAR_PLACES };
