--- src/mocks/offers.ts
@@ -1,421 +1,0 @@
-import { TOffer } from '../types/offer';
-
-const OFFERS: TOffer[] = [
-  {
-    id: '372dcd05-d179-4a2a-be70-0b08daada439',
-    title: 'Nice, cozy, warm big bed apartment',
-    type: 'room',
-    price: 167,
-    previewImage: 'https://13.design.pages.academy/static/hotel/3.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.36554,
-      longitude: 4.911976,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: true,
-    rating: 3.4,
-  },
-  {
-    id: '768226ff-11f3-4bec-b13c-a903de1da7de',
-    title: 'Perfectly located Castro',
-    type: 'room',
-    price: 253,
-    previewImage: 'https://13.design.pages.academy/static/hotel/9.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.385540000000006,
-      longitude: 4.902976,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.4,
-  },
-  {
-    id: 'f21b1d07-4ed8-456f-92ac-fd13b5cf6c19',
-    title: 'Loft Studio in the Central Area',
-    type: 'room',
-    price: 298,
-    previewImage: 'https://13.design.pages.academy/static/hotel/20.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.397540000000006,
-      longitude: 4.9099759999999995,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: true,
-    rating: 3.6,
-  },
-  {
-    id: '520aab03-3fbb-4528-8903-8f373247982b',
-    title: 'Tile House',
-    type: 'house',
-    price: 898,
-    previewImage: 'https://13.design.pages.academy/static/hotel/8.jpg',
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.37454,
-        longitude: 4.897976,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 52.37454,
-      longitude: 4.881976,
-      zoom: 16,
-    },
-    isFavorite: true,
-    isPremium: false,
-    rating: 1.9,
-  },
-  {
-    id: '068b3443-41f2-4d05-ac8d-6c2da2a96c9a',
-    title: 'The Pondhouse - A Magical Place',
-    type: 'house',
-    price: 429,
-    previewImage: 'https://13.design.pages.academy/static/hotel/17.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.868610000000004,
-      longitude: 2.342499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.4,
-  },
-  {
-    id: '413aa694-ca7b-4079-89f4-5a996f0c5da5',
-    title: 'Amazing and Extremely Central Flat',
-    type: 'apartment',
-    price: 188,
-    previewImage: 'https://13.design.pages.academy/static/hotel/8.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.858610000000006,
-      longitude: 2.330499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3.4,
-  },
-  {
-    id: '07dfea33-282a-404e-a9b9-77adc94262ed',
-    title: 'Perfectly located Castro',
-    type: 'house',
-    price: 166,
-    previewImage: 'https://13.design.pages.academy/static/hotel/17.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.834610000000005,
-      longitude: 2.335499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.8,
-  },
-  {
-    id: 'e74d855e-7a35-414a-911f-9c5d25aa413a',
-    title: 'Canal View Prinsengracht',
-    type: 'room',
-    price: 116,
-    previewImage: 'https://13.design.pages.academy/static/hotel/12.jpg',
-    city: {
-      name: 'Paris',
-      location: {
-        latitude: 48.85661,
-        longitude: 2.351499,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 48.85761,
-      longitude: 2.358499,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 4.3,
-  },
-
-  {
-    id: 'c6788ace-1dbd-45f0-8f7d-b5255f8e6af1',
-    title: 'Wood and stone place',
-    type: 'apartment',
-    price: 316,
-    previewImage: 'https://13.design.pages.academy/static/hotel/13.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.573341000000006,
-      longitude: 9.994654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.3,
-  },
-  {
-    id: '9a203b9e-25ae-4279-b1e2-c304e479a2da',
-    title: 'The Joshua Tree House',
-    type: 'room',
-    price: 287,
-    previewImage: 'https://13.design.pages.academy/static/hotel/15.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.528341000000005,
-      longitude: 9.982654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.5,
-  },
-  {
-    id: 'c720729a-2d7c-49fe-bcc7-1078d841a45a',
-    title: 'The house among olive ',
-    type: 'house',
-    price: 357,
-    previewImage: 'https://13.design.pages.academy/static/hotel/6.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.540341000000005,
-      longitude: 10.025654000000001,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3.7,
-  },
-  {
-    id: '6aaaa576-0853-4a12-adc2-31ddb30ae8ba',
-    title: 'The house among olive ',
-    type: 'apartment',
-    price: 273,
-    previewImage: 'https://13.design.pages.academy/static/hotel/14.jpg',
-    city: {
-      name: 'Hamburg',
-      location: {
-        latitude: 53.550341,
-        longitude: 10.000654,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 53.555341000000006,
-      longitude: 9.975654,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 3.9,
-  },
-  {
-    id: '626d04cc-a3f2-4732-8ced-051f55703ccd',
-    title: 'The Pondhouse - A Magical Place',
-    type: 'apartment',
-    price: 407,
-    previewImage: 'https://13.design.pages.academy/static/hotel/14.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.241402,
-      longitude: 6.782314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 1.4,
-  },
-  {
-    id: 'f4ac8b11-6b0e-438e-944d-8a725727b798',
-    title: 'The Pondhouse - A Magical Place',
-    type: 'room',
-    price: 109,
-    previewImage: 'https://13.design.pages.academy/static/hotel/17.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.235402,
-      longitude: 6.800314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 2.5,
-  },
-  {
-    id: '658dfd0d-e14b-4c2a-8f8d-c1573017c8a7',
-    title: 'Perfectly located Castro',
-    type: 'house',
-    price: 345,
-    previewImage: 'https://13.design.pages.academy/static/hotel/8.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.237402,
-      longitude: 6.779314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.8,
-  },
-  {
-    id: '0b2d7bb2-04c7-45f5-9f19-a6117997a8f1',
-    title: 'House in countryside',
-    type: 'hotel',
-    price: 397,
-    previewImage: 'https://13.design.pages.academy/static/hotel/2.jpg',
-    city: {
-      name: 'Dusseldorf',
-      location: {
-        latitude: 51.225402,
-        longitude: 6.776314,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 51.237402,
-      longitude: 6.797314,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 3,
-  },
-  {
-    id: '9be6e6f2-accf-481c-bc9e-db33328023c0',
-    title: 'Wood and stone place',
-    type: 'room',
-    price: 204,
-    previewImage: 'https://13.design.pages.academy/static/hotel/4.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.867557,
-      longitude: 4.371696999999999,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: true,
-    rating: 1.1,
-  },
-  {
-    id: '2e9817f6-d9eb-45e0-a207-bd0a8cd59855',
-    title: 'House in countryside',
-    type: 'room',
-    price: 276,
-    previewImage: 'https://13.design.pages.academy/static/hotel/18.jpg',
-    city: {
-      name: 'Brussels',
-      location: {
-        latitude: 50.846557,
-        longitude: 4.351697,
-        zoom: 13,
-      },
-    },
-    location: {
-      latitude: 50.827557,
-      longitude: 4.336697,
-      zoom: 16,
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 2.4,
-  },
-];
-
-export { OFFERS };
