--- src/mocks/reviews.ts
@@ -1,116 +1,0 @@
-import { TReview } from '../types/review';
-
-const REVIEWS: TReview[] = [
-  {
-    id: '372dcd05-d179-4a2a-be70-0b08daada439',
-    reviewComments: [
-      {
-        id: 'e64b03ed-f0a8-44c5-b20e-a991092a0dd7',
-        comment:
-          'I stayed here for one night and it was an unpleasant experience.',
-        date: '2023-06-29T21:00:00.490Z',
-        rating: 5,
-        user: {
-          name: 'Mollie',
-          avatarUrl: 'https://13.design.pages.academy/static/avatar/10.jpg',
-          isPro: false,
-        },
-      },
-      {
-        id: '185f44a5-a6ca-469c-aadd-a368531f67a4',
-        comment:
-          'Beautiful space, fantastic location and atmosphere, really a wonderful place to spend a few days. Will be back.',
-        date: '2023-06-29T21:00:00.490Z',
-        rating: 3,
-        user: {
-          name: 'Mollie',
-          avatarUrl: 'https://13.design.pages.academy/static/avatar/7.jpg',
-          isPro: true,
-        },
-      },
-      {
-        id: '794e803b-c626-49f7-8525-8ee56501014e',
-        comment:
-          'This villa is perfect in every way: the view on mountains and waterfalls, the hot tub and the villa itself. The evening here became a great continuation of our journeys over country.',
-        date: '2023-06-29T21:00:00.490Z',
-        rating: 5,
-        user: {
-          name: 'Isaac',
-          avatarUrl: 'https://13.design.pages.academy/static/avatar/10.jpg',
-          isPro: false,
-        },
-      },
-    ],
-  },
-  {
-    id: '768226ff-11f3-4bec-b13c-a903de1da7de',
-    reviewComments: [
-      {
-        id: '77135728-b48f-4b02-8f93-ce152a51ec1b',
-        comment:
-          'Beautiful space, fantastic location and atmosphere, really a wonderful place to spend a few days. Will be back.',
-        date: '2023-06-28T21:00:00.490Z',
-        rating: 1,
-        user: {
-          name: 'Emely',
-          avatarUrl: 'https://13.design.pages.academy/static/avatar/6.jpg',
-          isPro: false,
-        },
-      },
-      {
-        id: '6588c25d-d4f1-448f-82ec-e6797c6cb8e8',
-        comment:
-          'This villa is perfect in every way: the view on mountains and waterfalls, the hot tub and the villa itself. The evening here became a great continuation of our journeys over country.',
-        date: '2023-06-26T21:00:00.490Z',
-        rating: 4,
-        user: {
-          name: 'Sophie',
-          avatarUrl: 'https://13.design.pages.academy/static/avatar/8.jpg',
-          isPro: false,
-        },
-      },
-      {
-        id: '59b2964d-019c-426d-8652-c39e5216b377',
-        comment:
-          'This villa is perfect in every way: the view on mountains and waterfalls, the hot tub and the villa itself. The evening here became a great continuation of our journeys over country.',
-        date: '2023-06-23T21:00:00.490Z',
-        rating: 4,
-        user: {
-          name: 'Corey',
-          avatarUrl: 'https://13.design.pages.academy/static/avatar/7.jpg',
-          isPro: false,
-        },
-      },
-    ],
-  },
-  {
-    id: 'f21b1d07-4ed8-456f-92ac-fd13b5cf6c19',
-    reviewComments: [
-      {
-        id: 'c815d444-899c-4a5e-b150-e5834a1c8f2a',
-        comment:
-          'The house is very good, very happy, hygienic and simple living conditions around it are also very good. I hope to have the opportunity to come back. Thank you.',
-        date: '2023-06-29T21:00:00.490Z',
-        rating: 3,
-        user: {
-          name: 'Max',
-          avatarUrl: 'https://13.design.pages.academy/static/avatar/6.jpg',
-          isPro: true,
-        },
-      },
-      {
-        id: 'a3a9e293-b23f-4a11-9169-5a391441f1b8',
-        comment: 'Bathed in the nature. Completely unplugged. Unforgettable.',
-        date: '2023-06-27T21:00:00.490Z',
-        rating: 1,
-        user: {
-          name: 'Christina',
-          avatarUrl: 'https://13.design.pages.academy/static/avatar/2.jpg',
-          isPro: true,
-        },
-      },
-    ],
-  },
-];
-
-export { REVIEWS };
