+++ src/pages/favorites-page/favorites.tsx
@@ -32,7 +32,7 @@ const Favorites = () => {
     return <Spinner />;
   }
 
-  if (!favorites.length) {
+  if (loadingStatus === LoadingStatus.Error) {
     return <FavoritesEmpty />;
   }
